Added Working Testing framework
Added a working testing framework so that JUNIT Tests can be used. Tests should be placed in the tests folder, start with the name Test and end in .java. It should use standard JUnit tools. An example test, TestTest.Java has been created in order to test the Testing system.